renamed test folders
